test(advanced search): adapt tests (ongoing)
